[Algebra] Vector is now a specialization of Matrix.
This feature brings the semantic of the matrix and vector closer the
their mathematical definitions. Do to so, we use the Curiously recurring
template pattern (CRTP) in conjunction with partial specialization of
the class Matrix<R, C> where R is the number of rows and C is the number
of columns.

We now have the following structure:

BaseMatrix
   ├── Matrix<R, C> : General matrix of size RxC
   ├── BaseSquareMatrix
   │         ├─ Matrix<R, R> : Square matrix of dimensions RxR
   │         ├─ Matrix<2, 2> : Square matrix of dimensions 2x2
   │         └─ Matrix<3, 3> : Square matrix of dimensions 3x3
   └── BaseVector
             ├─Matrix<R, 1> : Row-vector matrix of dimensions Rx1
             ├─Matrix<1, C> : Col-vector matrix of dimensions 1xC
             ├─Matrix<1, 1> : Unit matrix of dimension 1x1
             └─BaseVector3D
                  ├─ Matrix<3, 1> : Row-vector matrix of dimensions 3x1
                  └─ Matrix<1, 3> : Col-vector matrix of dimensions 1x3

The user can than use Matrix<R,C> and the compiler will use the correct
specialization where some features are exclusively available (ex.
inverse is only available on square matrices).